Removed option-first restriction.
From second level onwards, commands should not use this options, as we want to be able to cascade options down arbitrary levels of commands.

We want `pyzure-cli <subcommand>... -h` not resolve to `pyzure-cli -h <subcommand>` which is different.

But we need still need the flag on the first level for options to be parsed properly, as was intended.