Lost third-level+ commands fixed.
In all our commands, we write the entire command chain in the usage pattern, which means we have to keep passing our entire command chain via `argv` to all our levels (below first, as that one gets it from `stdin` and has to cut itself out).

Which is what we're now doing in `storage` and `container` a second and third level command respectively.